+++ src/components/feedback/__tests__/feedback.test.js
@@ -1,11 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { testCases } from './feedback-test-cases.js';
-import Feedback from '..';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-configure({ adapter: new Adapter() });
 
 describe('feedback', () => {
   describe(testCases.basic.description, () => {
@@ -58,55 +53,5 @@ describe('feedback', () => {
     test('renders as expected', () => {
       expect(tree).toMatchSnapshot();
     });
-  });
-});
-
-describe('feedback interactions', () => {
-  let feedback;
-  beforeEach(() => {
-    feedback = mount(<Feedback site="dr-ui" location={{}} />);
-  });
-
-  test('character limit ok, user can submit feedback', () => {
-    feedback.setState({
-      helpful: true,
-      feedback: 'Lorem ipsum dolor sit amet'
-    });
-    // make sure submit button is not disabled
-    const submitBtn = feedback
-      .find('#dr-ui--feedback-submit-button')
-      .at(0)
-      .props();
-    expect(submitBtn.disabled).toBe(false);
-
-    // make sure overlimit warning is not shown
-    const overlimitWarning = feedback.find('#dr-ui--feedback-overlimit').at(0);
-    expect(overlimitWarning.exists()).toBe(false);
-
-    // make sure character counter is equal
-    const charCounter = feedback.find('#dr-ui--feedback-char-counter').at(0);
-    expect(charCounter.props().children).toEqual(974);
-  });
-
-  test('character limit too long, user cannot submit feedback', () => {
-    feedback.setState({
-      helpful: true,
-      feedback:
-        'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-    });
-    // make sure submit button is disabled
-    const submitBtn = feedback
-      .find('#dr-ui--feedback-submit-button')
-      .at(0)
-      .props();
-    expect(submitBtn.disabled).toBe(true);
-
-    // make sure overlimit warning is shown
-    const overlimitWarning = feedback.find('#dr-ui--feedback-overlimit').at(0);
-    expect(overlimitWarning.exists()).toBe(true);
-
-    // make sure character counter is equal
-    const charCounter = feedback.find('#dr-ui--feedback-char-counter').at(0);
-    expect(charCounter.props().children).toEqual(-340);
   });
 });
